pyproject.toml
[project]
name = "bnode-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Julius Aka", email = "julius.aka@uni-a.de" }
]
requires-python = ">=3.12"
dependencies = [
    "bokeh>=3.8.0",
    "dask>=2025.10.0",
    "distributed>=2025.10.0",
    "fmpy>=0.3.26",
    "h5py>=3.15.1",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.7",
    "mlflow>=3.5.0",
    "numpy>=2.3.4",
    "omegaconf>=2.3.0",
    "onnx>=1.19.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "pyyaml>=6.0.2",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "torch>=2.7.0",
    "torchaudio>=2.8.0",
    "torchdiffeq>=0.2.5",
    "torchvision>=0.22.0",
    "tqdm>=4.67.1",
]

[project.urls]
Homepage = " "
Documentation = " "
Repository = " "
Issues = " "

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "pytest-cov",
    "ty",
    "twine",
    "build",
    "pre-commit",
    "rich",
]

[project.scripts]
# this is a possibility to define commands
fmu_simulate = "bnode_core.data_generation.utils.fmu_simulate:main"
fmu_variable_names = "bnode_core.data_generation.utils.print_fmu_variable_names:main"
data_generation = "bnode_core.data_generation.raw_data_generation:main"
data_preperation = "bnode_core.data_generation.data_preperation:main"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py313"

[tool.coverage.run]
omit = ["tests/*"]
